backend/requirements.txt
# Backend requirements with exact versions
fastapi==0.114.2
python-multipart==0.0.7
email-validator==2.1.0.post1
passlib[bcrypt]==1.7.4
tenacity==8.2.3
pydantic==2.6.3
emails==0.6
jinja2==3.1.4
alembic==1.12.1
httpx==0.25.1
psycopg[binary]==3.1.13
sqlmodel==0.0.21
bcrypt==4.3.0
pydantic-settings==2.2.1
sentry-sdk[fastapi]==1.40.6
pyjwt==2.8.0
uvicorn==0.27.1

# Development dependencies
pytest==7.4.3
mypy==1.8.0
ruff==0.2.2
pre-commit==3.6.2
types-passlib==1.7.7.20240106
coverage==7.4.3
behave==1.2.6
pytest-bdd==7.0.1
pytest-cov==4.1.0